+++ examples/stress.rs
@@ -1,6 +1,3 @@
-extern crate pinboard;
-extern crate crossbeam;
-
 #[derive(Clone)]
 struct Test(u32);
 
@@ -22,7 +19,8 @@ fn main() {
                 }
             });
         }
-    }).unwrap();
+    })
+    .unwrap();
 
     println!("Exiting");
 }